requirements.txt
python-telegram-bot==20.7
SQLAlchemy==2.0.30
pymysql==1.1.0
